fix: refactor event route
